==> jupyter/0 Merge draw files into one.ipynb <==
import pandas as pd
import os

datapath = './data'

# 1 Join all Files together

def walk_through_files(path, file_extension=('.xlsx')):                # can check for both if tuple!
    for (dirpath, dirnames, filenames) in os.walk(path):
        for filename in filenames:
            if filename.endswith(file_extension): 
                yield os.path.join(dirpath, filename).replace("\\","/")        # ensure right backslashes (Windows)!
                
def generateList(directory):
    print("Generating List of documents in parent folder...")
    documentsList = []                            
    for fname in walk_through_files(directory):
        #print(fname)
        documentsList.append(fname)
    return documentsList

# Generate Dataset from individual Files

# mylist = generateList(datapath)
# finaldf = pd.DataFrame()
# for excelpart in mylist:
#     date = excelpart.split('/')[-1][:-5].replace('kino_','')           # retrieve year_month of file 
#     if date<'2010_11':           # after this date file format changes
#         df = pd.read_excel(excelpart, header = 1).sort_values(by=['Αριθμός Κλήρωσης'])   
#     else:
#         df = pd.read_excel(excelpart, header = 2).sort_values(by=['Αριθμός Κλήρωσης'])  
#     df = df[['Αριθμός Κλήρωσης', 'Ημ/νία Κλήρωσης', 'Ώρα Κλήρωσης', '1ος ', '2ος ',       # discard extra columns
#        '3ος ', '4ος ', '5ος ', '6ος ', '7ος ', '8ος ', '9ος ', '10ος ',
#        '11ος ', '12ος ', '13ος ', '14ος ', '15ος ', '16ος ', '17ος ', '18ος ',
#        '19ος ', '20ος ']]
#     finaldf = finaldf.append(df)
# finaldf.to_csv('./allDraws.csv')

# Load directly

finaldf = pd.read_csv('./allDraws.csv')
print(finaldf.shape)  
finaldf.head()
