==> jupyter/1 Feature Engineering.ipynb <==
# Notes

- Some ids are missing!

import pandas as pd
import numpy as np
import os
from numpy import random as nprandom
import random
import time

outputPath = os.path.join('..', 'features')

draws = pd.read_csv('../data/allDraws.csv')
draws = draws.rename(columns={'Αριθμός Κλήρωσης':'id', 'Ημ/νία Κλήρωσης':'date',
                             'Ώρα Κλήρωσης':'time', '1ος ':'WinNo1', '2ος ':'WinNo2', '3ος ':'WinNo3', '4ος ':'WinNo4',
                             '5ος ':'WinNo5', '6ος ':'WinNo6', '7ος ':'WinNo7', '8ος ':'WinNo8', '9ος ':'WinNo9', 
                              '10ος ':'WinNo10', '11ος ':'WinNo11', '12ος ':'WinNo12', '13ος ':'WinNo13', '14ος ':'WinNo14',
                              '15ος ':'WinNo15', '16ος ':'WinNo16', '17ος ':'WinNo17', '18ος ':'WinNo18', '19ος ':'WinNo19',
                              '20ος ':'WinNo20'})
draws.head()

# Targets

# Sort winning numbers (asc)

winCols = ['WinNo1', 'WinNo2', 'WinNo3', 'WinNo4', 'WinNo5', 'WinNo6', 'WinNo7', 'WinNo8', 'WinNo9', 'WinNo10',
             'WinNo11', 'WinNo12', 'WinNo13', 'WinNo14', 'WinNo15', 'WinNo16', 'WinNo17', 'WinNo18', 'WinNo19',
             'WinNo20']

winNumbers = draws[winCols].values
winNumbers.sort(axis=1) 
winNumbersDf = pd.DataFrame(winNumbers, draws.index, winCols)

# Rearrange Columns so id and winning Numbers can be seen as first columns

# for idx, winColName in enumerate(winCols):
#     winCol = winNumbersDf.pop(winColName)
#     winNumbersDf.insert(idx, winColName, winCol)

winNumbersDf.head()

# Datetime and Id features

# Extract time features

daysofweek = ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday']
monthsofyear = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August',
                'September', 'October', 'November', 'December']

featuresDf = draws[['id', 'date', 'time']].copy()
del draws

featuresDf['date'] = pd.to_datetime(featuresDf['date'])
featuresDf['time'] = pd.to_datetime(featuresDf['time'])
featuresDf['Hour'] = featuresDf['time'].dt.hour
featuresDf['Minutes'] = featuresDf['time'].dt.minute
featuresDf['Day'] = featuresDf['date'].dt.day
featuresDf['Month'] = featuresDf['date'].dt.month
featuresDf['Year'] = featuresDf['date'].dt.year
featuresDf.head()

# Prepare dictionaries

sinDoMDict28 = {}
for i in range(28):
    sinDoMDict28[i] = np.sin((2*np.pi*i)/28)
cosDoMDict28 = {}
for i in range(28):
    cosDoMDict28[i] = np.cos((2*np.pi*i)/28)
sinDoMDict30 = {}
for i in range(30):
    sinDoMDict30[i] = np.sin((2*np.pi*i)/30)
cosDoMDict30 = {}
for i in range(30):
    cosDoMDict30[i] = np.cos((2*np.pi*i)/30)
sinDoMDict31 = {}
for i in range(28):
    sinDoMDict31[i] = np.sin((2*np.pi*i)/31)
cosDoMDict31 = {}
for i in range(28):
    cosDoMDict31[i] = np.cos((2*np.pi*i)/31)

sindowDict = {}
for idx, day in enumerate(daysofweek):
    sindowDict[day]=np.sin((2*np.pi*idx)/len(daysofweek))

cosdowDict = {}
for idx, day in enumerate(daysofweek):
    cosdowDict[day]=np.cos((2*np.pi*idx)/len(daysofweek))

sinmoyDict = {}
for idx, month in enumerate(monthsofyear):
    sinmoyDict[month]=np.sin((2*np.pi*idx)/len(monthsofyear))

cosmoyDict = {}
for idx, month in enumerate(monthsofyear):
    cosmoyDict[month]=np.cos((2*np.pi*idx)/len(monthsofyear))

featuresDf['DayT'] = featuresDf['date'].dt.day_name()
featuresDf['MonthT'] = featuresDf['date'].dt.month_name()
    
# Hour of day
featuresDf['sinHod'] = np.round(np.sin((2*np.pi*(featuresDf['Hour'])/24)),7)
featuresDf['cosHod'] = np.round(np.cos((2*np.pi*(featuresDf['Hour'])/24)),7)

# Hour of half day (12h)
featuresDf['sinHohd'] = np.round(np.sin((2*np.pi*(featuresDf['Hour'])/12)),7)
featuresDf['cosHohd'] = np.round(np.cos((2*np.pi*(featuresDf['Hour'])/12)),7)

# Minute of Hour
featuresDf['sinMoH'] = np.round(np.sin((2*np.pi*(featuresDf['Minutes'])/60)),7)
featuresDf['cosMoH'] = np.round(np.cos((2*np.pi*(featuresDf['Minutes'])/60)),7)

# Minute of day
featuresDf['sinMoD'] = np.round(np.sin((2*np.pi*(featuresDf['Hour']*60+featuresDf['Minutes'])/1440)),7)
featuresDf['cosMoD'] = np.round(np.cos((2*np.pi*(featuresDf['Hour']*60+featuresDf['Minutes'])/1440)),7)

# DayName of week
featuresDf['sinDow'] = np.round(featuresDf['DayT'].map(sindowDict),7)              # angle represents periodic sin of Day of week!
featuresDf['cosDow'] = np.round(featuresDf['DayT'].map(cosdowDict),7)


# Day of Month
featuresDf['sinDoM28'] = np.round(featuresDf['Month'].map(sinDoMDict28),7)
featuresDf['cosDoM28'] = np.round(featuresDf['Month'].map(cosDoMDict28),7)
featuresDf['sinDoM30'] = np.round(featuresDf['Month'].map(sinDoMDict30),7)
featuresDf['cosDoM30'] = np.round(featuresDf['Month'].map(cosDoMDict30),7)
featuresDf['sinDoM31'] = np.round(featuresDf['Month'].map(sinDoMDict31),7)
featuresDf['cosDoM31'] = np.round(featuresDf['Month'].map(cosDoMDict31),7)

# Month of Year
featuresDf['sinMoy'] = featuresDf['MonthT'].map(sinmoyDict)
featuresDf['cosMoy'] = featuresDf['MonthT'].map(cosmoyDict)

# Year of Decade
featuresDf['sinYoD'] = np.round(np.sin((2*np.pi*(featuresDf['Year']%10)/10)),7)
featuresDf['cosYoD'] = np.round(np.cos((2*np.pi*(featuresDf['Year']%10)/10)),7)
featuresDf = featuresDf.drop(columns=['date', 'time', 'Hour', 'Minutes', 'Day', 'Month', 'DayT', 'MonthT', 'Year'])
print(featuresDf.columns)
featuresDf.head()

# Add exp and squared Transformations

relCols = [col for col in featuresDf.columns if col not in winCols and col!='id']

for col in relCols:
    featuresDf['e_'+col] = np.round(np.exp(featuresDf[col]),7)
    featuresDf['sq_'+col] = np.round(np.square(featuresDf[col]),7)     # when x in [-1,+1], log function is -inf

print(featuresDf.shape)
featuresDf.head()

# ContestId Features

primenumbers = [7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47]

start = time.time()

for primeno in primenumbers:
    featuresDf['sinId%'+str(primeno)] = np.round(np.sin(2*np.pi*(featuresDf['id']%primeno)/primeno),7)
    featuresDf['cosId%'+str(primeno)] = np.round(np.cos(2*np.pi*(featuresDf['id']%primeno)/primeno),7)

    featuresDf['e_sinId%'+str(primeno)] = np.round(np.exp(featuresDf['sinId%'+str(primeno)]),7)
    featuresDf['e_cosId%'+str(primeno)] = np.round(np.exp(featuresDf['cosId%'+str(primeno)]),7)

    featuresDf['sq_sinId%'+str(primeno)] = np.round(np.square(featuresDf['sinId%'+str(primeno)]),7)
    featuresDf['sq_cosId%'+str(primeno)] = np.round(np.square(featuresDf['cosId%'+str(primeno)]),7)

end = time.time()
elapsed = round(end-start,3)
print("Elapsed time: "+str(elapsed)+" seconds")
print(featuresDf.shape)
featuresDf.head(5)

# Merge

totaldf = winNumbersDf.join(featuresDf, how='outer')
idCol = totaldf.pop('id')
totaldf.insert(0, 'id', idCol)
print(totaldf.shape)
totaldf.head()

# Save features

if not os.path.exists(outputPath):
    os.mkdir(outputPath)
    
totaldf.to_csv(os.path.join(outputPath, "features.csv"), index=False)
totaldf.head()

# totaldf = pd.read_csv(os.path.join(outputPath, "featuresFull.csv"))
# small = totaldf.sample(frac=0.00303)
# small.shape
# small.to_csv(os.path.join(outputPath, "features.csv"), index=False)
# totaldf.shape
