==> jupyter/2 Batch Prep Modules.ipynb <==
from tensorflow.python.util import deprecation
deprecation._PRINT_DEPRECATION_WARNINGS = False

import pandas as pd
import numpy as np
import os
from numpy import random as nprandom
import random
import time
from multiprocessing import Pool
import sys
from queue import Queue
from threading import Thread
import multiprocessing
import tensorflow as tf

featuresPath = os.path.join('..', 'features')
lookback = 10
winNoCols = ['WinNo'+str(i) for i in range(1,21)]

df = pd.read_csv(os.path.join(featuresPath, 'features.csv'))
df.reset_index(drop=True, inplace=True)
assert not np.any(np.isnan(df.values))               # make sure we do not have nan values
df.head()

df2 = df[winNoCols]
df2.describe()

otherCols = [col for col in df.columns if col not in winNoCols and col!='id']
df3 = df[otherCols]
values = df3.values
print("Min value: ", values.min())
print("Max value: ", values.max())

# Prepare Dictionaries

numWinNumbers = 20
contFDict = {}
seqFDict = {}
winNumbersTDict = {}

start = time.time()
print("Creating dictionaries for batchPrep")
data = df.values

def getPastWinningNumbers(drawId): 
    return data[drawId-lookback:drawId, 1:numWinNumbers+1]
    
for example in [x for x in df.iloc[lookback:].values]:     # values start from id=lookback+1 !
    drawId = int(example[0])
    winNumbers = example[1:numWinNumbers+1]
    seqFDict[drawId] = getPastWinningNumbers(drawId)       # get previous 'lookback' times winning numbers
    contFDict[drawId] = example[numWinNumbers+1:]                       # continuous features
    winNumbersTDict[drawId] = winNumbers                   # the targets (12 numbers)
    
end = time.time()
elapsed = round(end-start,3)
print("Elapsed time: "+str(elapsed)+" seconds!")

# Batch Prep

batchSize = 64
batchesNo = 100
evalInterval = int(batchesNo/5)
drawIds = set(winNumbersTDict.keys())

def batchPrep():
    batchDrawIds = random.sample(drawIds, batchSize)
    contFeatures = np.array([contFDict[x] for x in batchDrawIds])
    seqFeatures = np.array([seqFDict[x] for x in batchDrawIds])
    targets = np.array([winNumbersTDict[x] for x in batchDrawIds])
    return contFeatures, seqFeatures, targets
    
start = time.time()
print("Running a batchprep simulation of size "+str(batchesNo))
for i in range(batchesNo):
    contFeatures, seqFeatures, targets = batchPrep()
    if i>0 and i%evalInterval==0:
        print("Eval time!")
        
end = time.time()
elapsed = round(end-start,3)
print("Elapsed time: "+str(elapsed)+" seconds!")
print("Batch speed: "+str(batchesNo/elapsed)+" batches/sec")

print(contFeatures.shape)       # batchsize, features
print(seqFeatures.shape)        # batchsize, timesteps (lookback), features
print(targets.shape)            # batchsize, targetsize

# 1st attempt: only feed features (no lstm)

## 1.2 Singlethreaded batchPrep

inputSize = contFeatures.shape[1]
print(inputSize)
targetSize = targets.shape[1]
print(targetSize)

# define weights and biases

tf.reset_default_graph()

biases = {
    'b1': tf.get_variable('b1', shape=(256,), initializer = tf.random_normal_initializer(0, 0.05)),
    'b2': tf.get_variable('b2', shape=(128,), initializer = tf.random_normal_initializer(0, 0.05)),
    'b3': tf.get_variable('b3', shape=(64,), initializer = tf.random_normal_initializer(0, 0.05)),
    'b4': tf.get_variable('b4', shape=(targetSize,), initializer = tf.random_normal_initializer(0, 0.05))
}

weights = {
    'w1': tf.get_variable('w1', shape=(inputSize, 256), initializer = tf.random_normal_initializer(0, 0.05)),
    'w2': tf.get_variable('w2', shape=(256, 128), initializer = tf.random_normal_initializer(0, 0.05)),
    'w3': tf.get_variable('w3', shape=(128, 64), initializer = tf.random_normal_initializer(0, 0.05)),
    'w4': tf.get_variable('w4', shape=(64, targetSize), initializer = tf.random_normal_initializer(0, 0.05))
}

# define TF layer functions and loss

def squared_loss(targets, prediction):
    return tf.reduce_mean(tf.math.abs(targets - prediction))

def fcLayer(x, w, b, activation=None):
    result = tf.matmul(x, w) + b
    if activation == 'relu':
        result = tf.nn.leaky_relu(result)         # normal will cut off most values
    if activation == 'softmax':
        result = tf.nn.softmax(result,axis=1)
    return result

# Graph operations

featuresInput = tf.placeholder(name="featuresInput", dtype=tf.float32, shape=(None, inputSize))
seqInput = ...
targetsOutput = tf.placeholder(name="targetsOutput", dtype=tf.float32, shape=(None, targetSize))

with tf.variable_scope("contFeatures"):
    x1 = fcLayer(featuresInput, weights['w1'], biases['b1'], 'relu')
    x2 = fcLayer(x1, weights['w2'], biases['b2'], 'relu')
    x3 = fcLayer(x2, weights['w3'], biases['b3'], 'relu')
    output = fcLayer(x3, weights['w4'], biases['b4']) 
# with tf.variable_scope("seqFeatures"):
#     y1 = 
    #y2 = 
# Concatenate
# Add a few more fcLayers
#with tf.variable_scope("jointFeatures"):
     

loss = squared_loss(targetsOutput, output)
optim = tf.train.AdamOptimizer(0.002, 0.5)
train = optim.minimize(loss)

config = tf.ConfigProto()
config.gpu_options.allow_growth = True
config.log_device_placement = False
config.allow_soft_placement = True

saver = tf.train.Saver()

def batchPrep():
    batchDrawIds = random.sample(drawIds, batchSize)
    contFeatures = np.array([contFDict[x] for x in batchDrawIds])
    seqFeatures = np.array([seqFDict[x] for x in batchDrawIds])
    targets = np.array([winNumbersTDict[x] for x in batchDrawIds])
    #targets2 = np.ones((64,20))
    return contFeatures, seqFeatures, targets

sess = tf.InteractiveSession(config=config)
sess.run(tf.global_variables_initializer())

from sklearn.metrics import mean_absolute_error
import itertools

random.seed(30)
evalDrawIds = random.sample(set(winNumbersTDict.keys()), int(len(df)*0.002))
testDrawIds = random.sample(set(winNumbersTDict.keys()), int(len(df)*0.002))
testDrawIds[:5]

# def processResults(targets, predictions):
    
maxPlayableNumbers = 12
winningNumbers = 20
poss = list(range(80))

np.random.seed(25)
targets = np.random.choice(poss, size=(2, winningNumbers), replace=False)
preds = np.random.choice(poss, size=(2, winningNumbers), replace=False)
preds.sort()
targets.sort()
print(targets)
print(preds)
#processResults(preds, targets)

set([1,3,3]).issubset(set([1,2,4,7,6,3,9]))

def processResults(targets, predictions):
    for i in range(len(targets)):
    #for predRow, targetRow in zip(predictions, targets):                   # iterate through rows
        targetSet = set(targets[i,:])
        predSet = set(predictions[i,:])
        #print(len(targetSet), len(predSet))
        for groupSize in range(10, maxPlayableNumbers+1):
            count=0
            recGroups = list(itertools.combinations(predSet, groupSize))
            #print(len(recGroups))
            for group in recGroups:
                if set(group).issubset(targetSet):      #set(group) <= targetSet:
                    count+=1
                    print("Count>0!")
                
start = time.time()

for i in range(20):
    processResults(targets, preds)
    
end = time.time()
elapsed = round(end-start,3)
print("Elapsed time: ",elapsed," seconds")

def processResults(targets, predictions):
    for predRow, targetRow in zip(predictions, targets):                   # iterate through rows
        targetSet = set(targetRow.flatten())
        predSet = set(predRow.flatten())
        print(len(targetSet), len(predSet))
        for groupSize in range(10, maxPlayableNumbers+1):
            count=0
            recGroups = list(itertools.combinations(predSet, groupSize))
            for group in recGroups:
                if group in targetSet:
                    count+=1
                    print("Count>0!")
            #print("\tGroupsize: ",groupSize, " valid Combinations: ", count," out of ",len(recGroups))


def runInference(runType):
    if runType=='eval':
        contFeatures = np.array([contFDict[x] for x in evalDrawIds])
        seqFeatures = np.array([seqFDict[x] for x in evalDrawIds])
        targets = np.array([winNumbersTDict[x] for x in evalDrawIds])
    elif runType=='test':
        contFeatures = np.array([contFDict[x] for x in testDrawIds])
        seqFeatures = np.array([seqFDict[x] for x in testDrawIds])
        targets = np.array([winNumbersTDict[x] for x in testDrawIds])
    else:
        raise Exception('runInference() runType should be in "eval" or "test", '+str(runType)+" provided")
    
    fd = {featuresInput: contFeatures, targetsOutput: targets}
    _, l, out = sess.run(fetches=[train, loss, output], feed_dict=fd)
    predictions = np.round(out)
    sqLoss = mean_absolute_error(targets, predictions)# ((targets - predictions)**2).mean()        # mean average error
    if runType=='eval':
        print("\tEval MAE: ", round(sqLoss,3))
    else:
        print("\tTest MAE: ", round(sqLoss,3))
        #processResults(targets, predictions)
        
    return predictions

batchSize = 100
batchesNo = 800

with tf.Session() as sess:
    sess.run(tf.global_variables_initializer())
    print("Training:")
    for i in range(batchesNo):
        contFeatures, seqFeatures, targets = batchPrep()
        #print("\tInput: ", contFeatures[0:1,0:5], ", targets:", targets[:1,:5])
        fd = {featuresInput: contFeatures, targetsOutput: targets}
        _, l, out, xx1, fi = sess.run(fetches=[train, loss, output, x1, featuresInput], feed_dict=fd)
        if i%(batchesNo/10)==0:
            print("\tBatch ",i,": Train loss: ", l)
        #print('\n\tOutput: ', out[1:3,1:3])
        #print("\tFeaturesInput: ", fi[0:1,0:5])
        #print("\tx1: ", xx1)
        if i>0 and i%100==0:
            runInference(runType='eval')
    print("Testing:")
    runInference(runType='test')        

# test

batchSize = 64
batchesNo = 1000
evalInterval = int(batchesNo/5)
drawIds = set(winNumbersTDict.keys())

def batchPrep():
    batchDrawIds = random.sample(drawIds, batchSize)
    contFeatures = np.array([contFDict[x] for x in batchDrawIds])
    seqFeatures = np.array([seqFDict[x] for x in batchDrawIds])
    targets = np.array([winNumbersTDict[x] for x in batchDrawIds])
    targets2 = np.ones((64,12))
    return contFeatures, seqFeatures, targets2
    
start = time.time()
print("Running a batchprep simulation of size "+str(batchesNo))
for i in range(batchesNo):
    contFeatures, seqFeatures, targets = batchPrep()
    if i>0 and i%evalInterval==0:
        print("Eval time!")
        
end = time.time()
elapsed = round(end-start,3)
print("Elapsed time: "+str(elapsed)+" seconds!")
print("Batch speed: "+str(batchesNo/elapsed)+" batches/sec")

## 1.2 Multithreaded Approach

a = [1,2,3]
b, c, g = a[0], a[1], a[2]

def batchPrep():
    """
    Prepares a batch and puts it in the queue for the other thread to process 
    """
    for i in range(batchesNo): 
        batchDrawIds = random.sample(drawIds, batchSize)                # prepares batch
        contFeatures = np.array([contFDict[x] for x in batchDrawIds])
        seqFeatures = np.array([seqFDict[x] for x in batchDrawIds])
        targets = np.array([winNumbersTDict[x] for x in batchDrawIds])
        q.put((contFeatures, seqFeatures, targets))                     # put it in the queue
    
def trainBatch(BatchTuple):
    """
    Feeds a batch to the model for training and prints Loss
    Args: 
      BatchTuple: the required input for the model corresponding to one batc/h 
    Returns: Sucess Code
    """
    # TF model train
    # print loss
    return 0

def runEvaluation():
    """
    Runs every batchesNo/5 steps to evaluate performance on eval set and prints loss
    Returns: Sucess Code
    """
    return 0

def trainModel(queue, batchesNo):
    """
    Spawns a thread to produce batches and main thread feeds them to the model
    Args: 
      queue: in OOP design, it will be self.queue
      batchesNo: in OOP design, it will be self.batchesNo
    """     
    t = Thread(target=batchPrep)
    t.daemon = True
    t.start()
    for i in range(batchesNo):  
        batchtuple = queue.get()
        trainBatch(batchtuple)
        if i>0 and i%(batchesNo/5)==0:
            print("Running Evaluation!")
            runEvaluation()
        queue.task_done()
    try:
        queue.join()                               # Join all Threads
    except KeyboardInterrupt:
        sys.exit(1)    

simNo = 1                      # number of simulations
start = time.time()
for i in range(simNo):            
    q = Queue(maxsize=5)       # batch queue
    batchesNo = 10000
    trainModel(q, batchesNo)                        

end = time.time()
elapsed = round(end-start,3)
print("Elapsed time: "+str(elapsed)+" seconds!")
if elapsed>0:
    print("Batch speed: "+str(simNo*batchesNo/elapsed)+" batches/sec")
